refactor: Search의 하위 컴포넌트로 SearchInput, SearchResult 생성, 모든 컴포넌트를 components 패키지로 이동
